requirements.txt
streamlit
google-generativeai
PyPDF2
langchain-community
duckduckgo-search
